#if UNITY_EDITOR でビルドエラーを修正
* PreviewSceneManager　が MonoBehaviour かつ Editor 専用なのが微妙
