+++ Assets/VRM10/Runtime/Components/Expression/Preview/PreviewMaterialUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45f9865ffd894b04bad0c39f53154f55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
